refactor: consolidate business features
